pod_capacity_dashboard/fapod/requirements.txt
asgiref
certifi
cffi
charset-normalizer
circuitbreaker
cryptography
Django==4.2.2
idna
oci
oci-cli
pycparser
pyOpenSSL
python-dateutil
python-dotenv
pytz
requests
six
sqlparse
urllib3
